pest-detection-fasstapi/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ML / YOLO
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# STT / TTS
faster-whisper>=1.0.0
openai>=1.30.0
requests>=2.31.0

# RAG / VectorDB
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
langchain-community>=0.0.36
langchain-text-splitters>=0.0.1
pypdf>=4.0.0

# Utils
numpy>=1.26.0
pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0